pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain-community>=0.3.21",
    "langchain>=0.3.24",
    "openai>=1.76.0",
    "prometheus-client>=0.21.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "langchain-openai>=0.3.14",
    "langchain-sqlserver>=0.1.1",
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "langchain-core>=0.3.56",
]
